+++ FirebaseMLModelDownloader/Sources/ModelDownloader.swift
@@ -22,6 +22,11 @@ import Foundation
 #endif // SWIFT_PACKAGE
 
 /// Possible ways to get a custom model.
+@available(
+  *,
+  deprecated,
+  message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+)
 public enum ModelDownloadType {
   /// Get local model stored on device if available. If no local model on device, this is the same
   /// as `latestModel`.
@@ -35,6 +40,11 @@ public enum ModelDownloadType {
 }
 
 /// Downloader to manage custom model downloads.
+@available(
+  *,
+  deprecated,
+  message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+)
 public class ModelDownloader {
   /// Name of the app associated with this instance of ModelDownloader.
   private let appName: String
@@ -99,6 +109,11 @@ public class ModelDownloader {
   }
 
   /// Model downloader with default app.
+  @available(
+    *,
+    deprecated,
+    message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+  )
   public static func modelDownloader() -> ModelDownloader {
     guard let defaultApp = FirebaseApp.app() else {
       fatalError(ModelDownloader.ErrorDescription.defaultAppNotConfigured)
@@ -107,6 +122,11 @@ public class ModelDownloader {
   }
 
   /// Model Downloader with custom app.
+  @available(
+    *,
+    deprecated,
+    message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+  )
   public static func modelDownloader(app: FirebaseApp) -> ModelDownloader {
     if let downloader = modelDownloaderDictionary[app.name] {
       DeviceLogger.logEvent(level: .debug,
@@ -133,6 +153,11 @@ public class ModelDownloader {
   /// download progress.
   ///   - completion: Returns either a `CustomModel` on success, or a `DownloadError` on failure, at
   /// the end of a model download.
+  @available(
+    *,
+    deprecated,
+    message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+  )
   public func getModel(name modelName: String,
                        downloadType: ModelDownloadType,
                        conditions: ModelDownloadConditions,
@@ -209,6 +234,11 @@ public class ModelDownloader {
   /// Gets the set of all downloaded models saved on device.
   /// - Parameter completion: Returns either a set of `CustomModel` models on success, or a
   /// `DownloadedModelError` on failure.
+  @available(
+    *,
+    deprecated,
+    message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+  )
   public func listDownloadedModels(completion: @escaping (Result<Set<CustomModel>,
     DownloadedModelError>) -> Void) {
     do {
@@ -274,6 +304,11 @@ public class ModelDownloader {
   ///   - modelName: The name of the model, matching Firebase console and already downloaded to
   /// device.
   ///   - completion: Returns a `DownloadedModelError` on failure.
+  @available(
+    *,
+    deprecated,
+    message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+  )
   public func deleteDownloadedModel(name modelName: String,
                                     completion: @escaping (Result<Void, DownloadedModelError>)
                                       -> Void) {
@@ -504,6 +539,11 @@ extension ModelDownloader {
 }
 
 /// Possible errors with model downloading.
+@available(
+  *,
+  deprecated,
+  message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+)
 public enum DownloadError: Error, Equatable {
   /// No model with this name exists on server.
   case notFound
@@ -524,6 +564,11 @@ public enum DownloadError: Error, Equatable {
 }
 
 /// Possible errors with locating a model file on device.
+@available(
+  *,
+  deprecated,
+  message: "FirebaseMLModelDownloader is deprecated and the service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml"
+)
 public enum DownloadedModelError: Error {
   /// No model with this name exists on device.
   case notFound