+++ FirebaseMLModelDownloader/CHANGELOG.md
@@ -1,2 +1,5 @@
+# Unreleased
+- [deprecated] FirebaseMLModelDownloader has been deprecated. The service will shut down on June 15, 2027. Use other custom model downloading alternatives. For more info, see https://firebase.google.com/docs/ml.
+
 # 12.8.0
 - [fixed] Remove unused legacy telemetry along with the large swift-protobuf dependency.